+++ ios/Classes/HeadlessInAppWebView/HeadlessInAppWebViewManager.swift
@@ -57,4 +57,15 @@ public class HeadlessInAppWebViewManager: NSObject, FlutterPlugin {
         headlessInAppWebView.onWebViewCreated()
         flutterWebView.makeInitialLoad(params: params as NSDictionary)
     }
+    
+    public func dispose() {
+        HeadlessInAppWebViewManager.channel?.setMethodCallHandler(nil)
+        HeadlessInAppWebViewManager.channel = nil
+        HeadlessInAppWebViewManager.registrar = nil
+        let headlessWebViews = HeadlessInAppWebViewManager.webViews.values
+        headlessWebViews.forEach { (headlessWebView: HeadlessInAppWebView) in
+            headlessWebView.dispose()
+        }
+        HeadlessInAppWebViewManager.webViews.removeAll()
+    }
 }