+++ ios/Classes/InAppWebViewMethodHandler.swift
@@ -598,8 +598,12 @@ public class InAppWebViewMethodHandler: FlutterMethodCallDelegate {
         }
     }
     
+    public func dispose() {
+        webView = nil
+    }
+    
     deinit {
         print("InAppWebViewMethodHandler - dealloc")
-        webView = nil
+        dispose()
     }
 }