+++ ios/Classes/SafariViewController/ChromeSafariBrowserManager.swift
@@ -91,4 +91,10 @@ public class ChromeSafariBrowserManager: NSObject, FlutterPlugin {
         
         result(FlutterError.init(code: "ChromeSafariBrowserManager", message: "SafariViewController is not available!", details: nil))
     }
+    
+    public func dispose() {
+        ChromeSafariBrowserManager.channel?.setMethodCallHandler(nil)
+        ChromeSafariBrowserManager.channel = nil
+        ChromeSafariBrowserManager.registrar = nil
+    }
 }