+++ ios/Classes/Types/WebMessage.swift
@@ -23,5 +23,6 @@ public class WebMessage : NSObject {
     
     deinit {
         print("WebMessage - dealloc")
+        dispose()
     }
 }